pyproject.toml
[tool.poetry]
name = "streamlit_dash"
version = "0.1.0"
description = ""
authors = ["GustavoDiez <gustavo@nirakara.org>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
jupyter = "^1.0.0"
numpy = "^1.22.1"
streamlit = "^1.4.0"
google-cloud-bigquery = "^2.32.0"
google-cloud = "^0.34.0"
plotly-express = "^0.4.1"
google-cloud-storage = "^2.1.0"
altair = "^4.2.0"
vega-datasets = "^0.9.0"
altair-catplot = "^0.0.5"
watchdog = "^2.1.6"
plotly = "^5.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
